+++ PUBG/Client/Source/UI/UIObject.cpp
@@ -42,7 +42,11 @@ void UIObject::DrawBorder()
     
     const auto d = Device()();
     d->SetFVF(VERTEX_RHWC::FVF);
-    d->DrawPrimitiveUP(D3DPT_LINESTRIP, vertices.size() - 1, vertices.data(), sizeof VERTEX_RHWC);
+    d->DrawPrimitiveUP(
+        D3DPT_LINESTRIP, 
+        static_cast<UINT>(vertices.size() - 1), 
+        vertices.data(), 
+        sizeof VERTEX_RHWC);
 }
 
 void UIObject::SetViewportPosition(const D3DXVECTOR3& parentViewportPos, const D3DXVECTOR3& pos)
