+++ PUBG/Client/Source/UI/UIText.cpp
@@ -39,11 +39,23 @@ void UIText::Render()
 
     if (m_pText)
     {
-        m_pFont->DrawTextA(Sprite()(), m_pText->c_str(), m_pText->size(), &m_rect, m_drawTextFormat, m_color);
+        m_pFont->DrawTextA(
+            Sprite()(), 
+            m_pText->c_str(), 
+            static_cast<INT>(m_pText->size()), 
+            &m_rect, 
+            m_drawTextFormat, 
+            m_color);
     }
     else
     {
-        m_pFont->DrawTextA(Sprite()(), m_text.c_str(), m_text.size(), &m_rect, m_drawTextFormat, m_color);
+        m_pFont->DrawTextA(
+            Sprite()(), 
+            m_text.c_str(), 
+            static_cast<INT>(m_text.size()),
+            &m_rect, 
+            m_drawTextFormat, 
+            m_color);
     }
     
     UIObject::Render();
